requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
Flask-SQLAlchemy==3.1.1
requests==2.32.5
python-dotenv==1.2.1
gunicorn==23.0.0
SQLAlchemy==2.0.44
